refactor: refactor callbacks to promises
Refactored tests and implementation to promises.
